refactor(analysis-types): extract effort model DTO
